+++ CHANGELOG.rst
@@ -7,6 +7,13 @@ What's new
 All notable changes to the codebase are documented in this file.
 
 
+Version 1.3 (2025-06-27)
+--------------------------
+- Fixes to the pair-matching algorithm within the sexual network to better align partner ages 
+- Improvements to networks, including analyzers for debut age and partner age differences
+- *GitHub info*: PR `143 <https://github.com/starsimhub/stisim/pull/143>`_
+
+
 Version 1.2 (2025-06-10)
 --------------------------
 - Improvements to networks, including analyzers for relationship duration and network degree